+++ Demo/DemoTests/MoyaProviderIntegrationTests.swift
@@ -1,9 +1,59 @@
 import UIKit
-import Moya
 import Quick
 import Nimble
 import ReactiveMoya
 import OHHTTPStubs
+
+private extension String {
+    var URLEscapedString: String {
+        return self.stringByAddingPercentEncodingWithAllowedCharacters(NSCharacterSet.URLHostAllowedCharacterSet())!
+    }
+}
+
+private enum GitHub {
+    case Zen
+    case UserProfile(String)
+}
+
+extension GitHub : MoyaTarget {
+    var baseURL: NSURL { return NSURL(string: "https://api.github.com")! }
+    var path: String {
+        switch self {
+        case .Zen:
+            return "/zen"
+        case .UserProfile(let name):
+            return "/users/\(name.URLEscapedString)"
+        }
+    }
+    var method: Moya.Method {
+        return .GET
+    }
+    var parameters: [String: AnyObject] {
+        return [:]
+    }
+    var sampleData: NSData {
+        switch self {
+        case .Zen:
+            return "Half measures are as bad as nothing at all.".dataUsingEncoding(NSUTF8StringEncoding)!
+        case .UserProfile(let name):
+            return "{\"login\": \"\(name)\", \"id\": 100}".dataUsingEncoding(NSUTF8StringEncoding)!
+        }
+    }
+}
+
+private func url(route: MoyaTarget) -> String {
+    return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
+}
+
+private let lazyEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Closure({.Success(200, {target.sampleData})}), method: target.method, parameters: target.parameters)
+}
+
+private let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    let errorData = "Houston, we have a problem".dataUsingEncoding(NSUTF8StringEncoding)!
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Error(401, NSError(domain: "com.moya.error", code: 0, userInfo: nil), {errorData}), method: target.method, parameters: target.parameters)
+}
+
 
 func beIndenticalToResponse(expectedValue: MoyaResponse) -> MatcherFunc<MoyaResponse> {
     return MatcherFunc { actualExpression, failureMessage in
