+++ Demo/DemoTests/MoyaProviderSpec.swift
@@ -1,11 +1,7 @@
 import Quick
-import Moya
-import RxMoya
-import ReactiveMoya
 import Nimble
-import ReactiveCocoa
-import RxSwift
 import Alamofire
+import Moya
 
 class MoyaProviderSpec: QuickSpec {
     override func spec() {
@@ -47,7 +43,7 @@ class MoyaProviderSpec: QuickSpec {
             
             let endpoint1 = provider.endpoint(target)
             let endpoint2 = provider.endpoint(target)
-            expect(endpoint1).to(equal(endpoint2))
+            expect(endpoint1.urlRequest).to(equal(endpoint2.urlRequest))
         }
         
         it("returns a cancellable object when a request is made") {
@@ -207,4 +203,54 @@ class MoyaProviderSpec: QuickSpec {
             }
         }
     }
+}
+
+private extension String {
+    var URLEscapedString: String {
+        return self.stringByAddingPercentEncodingWithAllowedCharacters(NSCharacterSet.URLHostAllowedCharacterSet())!
+    }
+}
+
+private enum GitHub {
+    case Zen
+    case UserProfile(String)
+}
+
+extension GitHub : MoyaTarget {
+    var baseURL: NSURL { return NSURL(string: "https://api.github.com")! }
+    var path: String {
+        switch self {
+        case .Zen:
+            return "/zen"
+        case .UserProfile(let name):
+            return "/users/\(name.URLEscapedString)"
+        }
+    }
+    var method: Moya.Method {
+        return .GET
+    }
+    var parameters: [String: AnyObject] {
+        return [:]
+    }
+    var sampleData: NSData {
+        switch self {
+        case .Zen:
+            return "Half measures are as bad as nothing at all.".dataUsingEncoding(NSUTF8StringEncoding)!
+        case .UserProfile(let name):
+            return "{\"login\": \"\(name)\", \"id\": 100}".dataUsingEncoding(NSUTF8StringEncoding)!
+        }
+    }
+}
+
+private func url(route: MoyaTarget) -> String {
+    return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
+}
+
+private let lazyEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Closure({.Success(200, {target.sampleData})}), method: target.method, parameters: target.parameters)
+}
+
+private let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    let errorData = "Houston, we have a problem".dataUsingEncoding(NSUTF8StringEncoding)!
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Error(401, NSError(domain: "com.moya.error", code: 0, userInfo: nil), {errorData}), method: target.method, parameters: target.parameters)
 }
