+++ Demo/DemoTests/ReactiveCocoaMoyaProviderTests.swift
@@ -1,8 +1,7 @@
 import Quick
 import Nimble
-import Moya
-import ReactiveMoya
 import ReactiveCocoa
+import ReactiveMoya
 import Alamofire
 
 class ReactiveCocoaMoyaProviderSpec: QuickSpec {
@@ -147,4 +146,54 @@ class ReactiveCocoaMoyaProviderSpec: QuickSpec {
             }
         }
     }
+}
+
+private extension String {
+    var URLEscapedString: String {
+        return self.stringByAddingPercentEncodingWithAllowedCharacters(NSCharacterSet.URLHostAllowedCharacterSet())!
+    }
+}
+
+private enum GitHub {
+    case Zen
+    case UserProfile(String)
+}
+
+extension GitHub : MoyaTarget {
+    var baseURL: NSURL { return NSURL(string: "https://api.github.com")! }
+    var path: String {
+        switch self {
+        case .Zen:
+            return "/zen"
+        case .UserProfile(let name):
+            return "/users/\(name.URLEscapedString)"
+        }
+    }
+    var method: Moya.Method {
+        return .GET
+    }
+    var parameters: [String: AnyObject] {
+        return [:]
+    }
+    var sampleData: NSData {
+        switch self {
+        case .Zen:
+            return "Half measures are as bad as nothing at all.".dataUsingEncoding(NSUTF8StringEncoding)!
+        case .UserProfile(let name):
+            return "{\"login\": \"\(name)\", \"id\": 100}".dataUsingEncoding(NSUTF8StringEncoding)!
+        }
+    }
+}
+
+private func url(route: MoyaTarget) -> String {
+    return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
+}
+
+private let lazyEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Closure({.Success(200, {target.sampleData})}), method: target.method, parameters: target.parameters)
+}
+
+private let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
+    let errorData = "Houston, we have a problem".dataUsingEncoding(NSUTF8StringEncoding)!
+    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Error(401, NSError(domain: "com.moya.error", code: 0, userInfo: nil), {errorData}), method: target.method, parameters: target.parameters)
 }
