--- Demo/DemoTests/TestResources.swift
@@ -1,53 +1,0 @@
-import Foundation
-import Moya
-import UIKit
-
-private extension String {
-    var URLEscapedString: String {
-        return self.stringByAddingPercentEncodingWithAllowedCharacters(NSCharacterSet.URLHostAllowedCharacterSet())!
-    }
-}
-
-enum GitHub {
-    case Zen
-    case UserProfile(String)
-}
-
-extension GitHub : MoyaTarget {
-    var baseURL: NSURL { return NSURL(string: "https://api.github.com")! }
-    var path: String {
-        switch self {
-        case .Zen:
-            return "/zen"
-        case .UserProfile(let name):
-            return "/users/\(name.URLEscapedString)"
-        }
-    }
-    var method: Moya.Method {
-        return .GET
-    }
-    var parameters: [String: AnyObject] {
-        return [:]
-    }
-    var sampleData: NSData {
-        switch self {
-        case .Zen:
-            return "Half measures are as bad as nothing at all.".dataUsingEncoding(NSUTF8StringEncoding)!
-        case .UserProfile(let name):
-            return "{\"login\": \"\(name)\", \"id\": 100}".dataUsingEncoding(NSUTF8StringEncoding)!
-        }
-    }
-}
-
-public func url(route: MoyaTarget) -> String {
-    return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
-}
-
-let lazyEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
-    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Closure({.Success(200, {target.sampleData})}), method: target.method, parameters: target.parameters)
-}
-
-let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
-    let errorData = "Houston, we have a problem".dataUsingEncoding(NSUTF8StringEncoding)!
-    return Endpoint<GitHub>(URL: url(target), sampleResponse: .Error(401, NSError(domain: "com.moya.error", code: 0, userInfo: nil), {errorData}), method: target.method, parameters: target.parameters)
-}
