+++ Moya/ReactiveCocoa/Moya+ReactiveCocoa.swift
@@ -1,5 +1,4 @@
 import Foundation
-import Moya
 import ReactiveCocoa
 import Alamofire
