+++ Moya/RxSwift/Moya+RxSwift.swift
@@ -1,5 +1,4 @@
 import Foundation
-import Moya
 import RxSwift
 import Alamofire
